include DevSupport
Without this the dev menu will not appear, after the shake gesture.
I don't know if we need that, but here there are also some flags that
are adjusted in the solution proposed here:
https://github.com/facebook/react-native/issues/2133
Note that those flags are not required for the dev menu to work, so I've
skipped them.